requirements.txt
typedb-driver==3.8.1